uniformity of docs page

+++ invite_only_app/lib/widgets/pages/docs_page.dart
@@ -5,6 +5,7 @@ import 'package:invite_only_app/blocs/auth/auth_state.dart';
 import 'package:invite_only_app/blocs/docs/docs_bloc.dart';
 import 'package:invite_only_app/blocs/docs/docs_event.dart';
 import 'package:invite_only_app/blocs/docs/docs_state.dart';
+import 'package:invite_only_app/widgets/dialogs/confirmation_dialog.dart';
 import 'package:invite_only_app/widgets/other/error_message.dart';
 import 'package:invite_only_repo/invite_only_repo.dart';
 import 'package:rsa_scan/rsa_scan.dart';
@@ -17,7 +18,7 @@ class DocsPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(title: Text('My Documents')),
+      appBar: AppBar(title: Text('My Profile')),
       body: BlocBuilder<DocsBloc, DocsState>(
         builder: (context, state) {
           if (state is LoadingDocs) {
@@ -97,6 +98,13 @@ class DocsPage extends StatelessWidget {
     if (T == Passport) uploaded = state.passport != null;
 
     return ListTile(
+      leading: Builder(builder: (context) {
+        if (T == IdCard) return Icon(Icons.credit_card);
+        if (T == IdBook) return Icon(Icons.book);
+        if (T == DriversLicense) return Icon(Icons.drive_eta);
+        if (T == Passport) return Icon(Icons.airplanemode_active);
+        return null;
+      }),
       title: Builder(builder: (context) {
         if (T == IdCard) return Text('ID Card');
         if (T == IdBook) return Text('ID Book');
@@ -105,45 +113,47 @@ class DocsPage extends StatelessWidget {
         return null;
       }),
       subtitle: Text(
-        !uploaded ? "Not Uploaded" : "Uploaded",
+        uploaded ? 'Remove the existing document' : 'Scan a new document',
       ),
-      trailing: Row(
-        mainAxisSize: MainAxisSize.min,
-        children: !uploaded
-            ? [
-                IconButton(
-                  icon: Icon(Icons.camera_alt),
-                  onPressed: () async {
-                    var rsa;
-                    if (T == IdCard) rsa = await scanIdCard(context);
-                    if (T == IdBook) rsa = await scanIdBook(context);
-                    if (T == DriversLicense) rsa = await scanDrivers(context);
-                    if (T == Passport) rsa = await scanId(context);
+      trailing: Builder(builder: (context) {
+        if (uploaded) {
+          return IconButton(
+            icon: Icon(Icons.delete),
+            onPressed: () async {
+              showConfirmationDialog(
+                context,
+                'Deleting this document will remove it from your account.',
+                () {
+                  if (T == IdCard)
+                    DocsBloc.of(context).add(DeleteDoc(state.idCard));
+                  if (T == IdBook)
+                    DocsBloc.of(context).add(DeleteDoc(state.idBook));
+                  if (T == DriversLicense)
+                    DocsBloc.of(context).add(DeleteDoc(state.driversLicense));
+                  if (T == Passport)
+                    DocsBloc.of(context).add(DeleteDoc(state.passport));
+                },
+              );
+            },
+          );
+        } else {
+          return IconButton(
+            icon: Icon(Icons.camera_alt),
+            onPressed: () async {
+              var rsa;
+              if (T == IdCard) rsa = await scanIdCard(context);
+              if (T == IdBook) rsa = await scanIdBook(context);
+              if (T == DriversLicense) rsa = await scanDrivers(context);
+              if (T == Passport) rsa = await scanId(context);
 
-                    if (rsa == null) return;
-                    DocsBloc.of(context).add(
-                      SubmitDoc(DocsBloc.rsaToDocs(rsa)),
-                    );
-                  },
-                ),
-              ]
-            : [
-                Icon(Icons.check_circle, color: Colors.green),
-                IconButton(
-                  icon: Icon(Icons.delete),
-                  onPressed: () async {
-                    if (T == IdCard)
-                      DocsBloc.of(context).add(DeleteDoc(state.idCard));
-                    if (T == IdBook)
-                      DocsBloc.of(context).add(DeleteDoc(state.idBook));
-                    if (T == DriversLicense)
-                      DocsBloc.of(context).add(DeleteDoc(state.driversLicense));
-                    if (T == Passport)
-                      DocsBloc.of(context).add(DeleteDoc(state.passport));
-                  },
-                ),
-              ],
-      ),
+              if (rsa == null) return;
+              DocsBloc.of(context).add(
+                SubmitDoc(DocsBloc.rsaToDocs(rsa)),
+              );
+            },
+          );
+        }
+      }),
     );
   }
 }
